% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{layer_image_resize}
\alias{layer_image_resize}
\title{Resize Images}
\usage{
layer_image_resize(object, size, align_corners = FALSE, method = "bilinear",
  name = NULL)
}
\arguments{
\item{object}{Model or layer object}

\item{size}{integer of 2 elements: new_height, new_width. The new size for the images}

\item{align_corners}{If TRUE, the centers of the 4 corner pixels of the
input and output tensors are aligned, preserving the values at the corner
pixels. Defaults to FALSE.}

\item{method}{ResizeMethod. Defaults to "bilinear".}

\item{name}{An optional name string for the layer. Should be unique in a
model (do not reuse the same name twice). It will be autogenerated if it
isn't provided.}
}
\description{
Resize Images
}
\examples{

\dontrun{
images <- layer_input(c(200, 200, 3))
images_resized <- layer_image_resize(images, c(100, 100))
}

}
\seealso{
Other image: \code{\link{layer_image_rgb_to_grayscale}}
}
