% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{layer_rgb_to_grayscale}
\alias{layer_rgb_to_grayscale}
\title{RGB to Grayscale}
\usage{
layer_rgb_to_grayscale(object, name = NULL)
}
\arguments{
\item{object}{Model or layer object}

\item{name}{An optional name string for the layer. Should be unique in a
model (do not reuse the same name twice). It will be autogenerated if it
isn't provided.}
}
\description{
Converts an image in RGB to the Grayscale.
}
\examples{
\dontrun{
rgb_images <- layer_input(c(200, 200, 3))
gs_images <- layer_rgb_to_grayscale(rgb_images)
}

}
