% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrogram.R
\name{layer_spectrogram}
\alias{layer_spectrogram}
\title{Layer Spectrogram}
\usage{
layer_spectrogram(object, frame_length, frame_step, fft_length = NULL,
  pad_end = FALSE, name = NULL)
}
\arguments{
\item{object}{Model or layer object}

\item{frame_length}{The window length in samples.}

\item{frame_step}{The number of samples to step.}

\item{fft_length}{The size of the FFT to apply. If not provided, uses the
smallest power of 2 enclosing frame_length.}

\item{pad_end}{Whether to pad the end of signals with zeros when the provided
frame length and step produces a frame that lies partially past its end.}

\item{name}{An optional name string for the layer. Should be unique in a
model (do not reuse the same name twice). It will be autogenerated if it
isn't provided.}
}
\description{
Computes the spectrogram of a signal using the STFT implemented in
tf.contrib.signal.
}
